major refactoring with new package name
