+++ setup.py
@@ -5,12 +5,12 @@ requirements = []
 test_requirements = ['pytest>=3']
 
 setup(
-    name='sinequa_py',
+    name='pynequa',
     version=VERSION,
     author='Anish Bhusal',
     author_email='anish.bhusal@uah.edu',
     description='Pythonic Layer for Sinequa REST API',
-    packages=find_packages(include=['sinequa_py', 'sinequa_py.*']),
+    packages=find_packages(include=['pynequa', 'pynequa.*']),
     install_requires=requirements,
     test_suite='tests',
     tests_require=test_requirements,
